# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jack — AI Trade Intelligence Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List all videos in the knowledge library
      parameters:
        - name: trade
          in: query
          schema:
            type: string
          description: Filter by trade category
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, uploading, uploaded, transcribing, analyzing, indexing, completed, failed, retrying]
          description: Filter by processing status
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoList"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Register a new video upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
...
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a single video with its transcript and analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Video detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoDetail"
        "404":
          description: Not found
    patch:
      operationId: updateVideo
      tags: [videos]
      summary: Update video metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoUpdate"
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video and all its data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /videos/{id}/transcribe:
    post:
      operationId: transcribeVideo
      tags: [videos]
      summary: Trigger AI transcription via Whisper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Transcription queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
  /videos/{id}/analyze:
    post:
      operationId: analyzeVideo
      tags: [videos]
      summary: Trigger GPT-4 analysis and Red Seal competency mapping
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Analysis queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
  /videos/{id}/related:
    get:
      operationId: fetchRelatedVideos
      tags: [videos]
      summary: Get semantically related videos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Related videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
  /videos/{id}/upload-url:
    post:
      operationId: getUploadUrl
      tags: [videos]
      summary: Get a signed Supabase Storage upload URL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Signed upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
...
components:
  schemas:
...
    Video:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        trade:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        duration:
          type: ["number", "null"]
        status:
          type: string
          enum: [queued, uploading, uploaded, transcribing, analyzing, indexing, completed, failed, retrying]
        competencyCodes:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        attempts:
          type: ["integer", "null"]
        lastError:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    VideoDetail:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        trade:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        duration:
          type: ["number", "null"]
        status:
          type: string
          enum: [queued, uploading, uploaded, transcribing, analyzing, indexing, completed, failed, retrying]
        competencyCodes:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        attempts:
          type: ["integer", "null"]
        lastError:
          type: ["string", "null"]
        transcript:
          type: ["string", "null"]
        segments:
          type: array
          items:
            $ref: "#/components/schemas/TranscriptSegment"
        analysis:
          type: ["string", "null"]
        keyPoints:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    VideoList:
      type: object
      required: [videos, total]
      properties:
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        total:
          type: integer
    VideoInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        trade:
          type: string
        tags:
          type: array
          items:
            type: string
    VideoUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        trade:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [queued, uploading, uploaded, transcribing, analyzing, indexing, completed, failed, retrying]
...
    TranscriptSegment:
      type: object
      required: [id, startTime, endTime, text]
      properties:
        id:
          type: string
        startTime:
          type: number
        endTime:
          type: number
        text:
          type: string
        confidence:
          type: ["number", "null"]
    JobStatus:
      type: object
      required: [jobId, status, videoId]
      properties:
        jobId:
          type: string
        status:
          type: string
        videoId:
          type: string
        message:
          type: ["string", "null"]
...
    UploadUrlRequest:
      type: object
      required: [filename, contentType]
      properties:
        filename:
          type: string
        contentType:
          type: string
    UploadUrlResponse:
      type: object
      required: [uploadUrl, path]
      properties:
        uploadUrl:
          type: string
        path:
          type: string
        token:
          type: ["string", "null"]
...